Make form, model and controller distinct classes
